requirements.txt
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==3.2.13
djangorestframework==3.13.1
filelock==3.4.1
idna==3.3
importlib-metadata==4.8.3
kafka==1.3.5
packaging==21.3
pkg-resources==0.0.0
pyparsing==3.0.8
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.1.2
typing-extensions==4.1.1
urllib3==1.26.9
wrapt==1.14.0
zipp==3.6.0
